requirements.txt
python-dotenv>=1.0.0

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0

openai>=1.0.0
pinecone>=5.0.0
pypdf>=3.17.0
docx2txt>=0.8
numpy>=1.24.0

fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

streamlit>=1.30.0
requests>=2.31.0
aiohttp>=3.9.0

# Document generation
python-docx>=1.1.0

# OCR support (PyMuPDF - no poppler needed!)
pymupdf>=1.24.0
Pillow>=10.0.0
gdown>=5.1.0
